\name{fsh}
\alias{fsh}

\title{ Flush Console Buffer }
\description{
  Flush contents of console buffer.  This function only has an effect on
  the Windows GUI. 
}
\usage{
fsh()
}
\value{
  No value is returned.
}

\author{ Peter Rossi, Graduate School of Business, University of Chicago, 
  \email{Peter.Rossi@ChicagoGsb.edu}.
}

\keyword{ utilities }
