\name{fsh}
\alias{fsh}

\title{ Flush Console Buffer }
\description{
  Flush contents of console buffer.  This function only has an effect on
  the Windows GUI. 
}
\usage{
fsh()
}
\value{
  No value is returned.
}

\author{ Peter Rossi, Anderson School, UCLA, 
  \email{perossichi@gmail.com}.
}

\keyword{ utilities }
